# API-Spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Employee Task Management REST API",
    "version": "1.0.0",
    "description": "REST API documentation for Employee Task Management application with JWT security."
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/tasks/{id}": {
      "get": {
        "summary": "Get task by ID",
        "tags": [
          "task-controller"
        ],
        "operationId": "getTaskById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                },
                "examples": {
                  "getTaskByIdSuccess": {
                    "summary": "Get Task By ID - Success",
                    "value": {
                      "id": 1,
                      "title": "Implement Login API",
                      "description": "Create login endpoint using JWT",
                      "status": "PENDING",
                      "priority": "HIGH",
                      "dueDate": "2026-08-10",
                      "assignedUserId": 17,
                      "assignedUserName": "zohaib"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "getTaskByIdAuthenticationFailed": {
                    "summary": "Get Task By ID - Authentication Failed",
                    "value": {
                      "message": "Authentication failed",
                      "status": 401,
                      "timestamp": "2026-08-07T10:14:22.04642"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Requires ADMIN role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "getTaskByIdAccessDenied": {
                    "summary": "Get Task By ID - Access Denied",
                    "value": {
                      "message": "Access denied. You do not have permission to access this resource.",
                      "status": 403,
                      "timestamp": "2026-08-07T10:10:04.673576"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Task not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "getTaskByIdNotFound": {
                    "summary": "Get Task By ID - Task Not Found",
                    "value": {
                      "message": "Task not found",
                      "status": 404,
                      "timestamp": "2026-08-07T12:44:43.081885"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update task details",
        "tags": [
          "task-controller"
        ],
        "operationId": "updateTask",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                },
                "examples": {
                  "updateTaskSuccessAdminToken": {
                    "summary": "Update Task - Success Admin token",
                    "value": {
                      "id": 3,
                      "title": "Complete Spring Security",
                      "description": "Implement JWT authentication",
                      "status": "PENDING",
                      "priority": "HIGH",
                      "dueDate": "2026-08-15",
                      "assignedUserId": 18,
                      "assignedUserName": "amash"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Validation error or invalid priority/status enum",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "updateTaskCannotAssignToAdmin": {
                    "summary": "Update Task - Cannot Assign To Admin",
                    "value": {
                      "message": "Task can only be assigned to employees",
                      "status": 400,
                      "timestamp": "2026-08-07T11:58:12.019669"
                    }
                  },
                  "updateTaskInvalidPriority": {
                    "summary": "Update Task - Invalid Priority",
                    "value": {
                      "message": "Invalid priority. Allowed values: HIGH, LOW, MEDIUM",
                      "status": 400,
                      "timestamp": "2026-08-07T11:58:27.27979"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "updateTaskNoToken": {
                    "summary": "Update Task - No token",
                    "value": {
                      "message": "Authentication failed",
                      "status": 401,
                      "timestamp": "2026-08-07T11:57:24.764987"
                    }
                  },
                  "updateTaskInvalidToken": {
                    "summary": "Update Task - Invalid token",
                    "value": {
                      "message": "Authentication failed",
                      "status": 401,
                      "timestamp": "2026-08-07T11:57:38.222514"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Requires ADMIN role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "updateTaskEmployeeTokenAccessDenied": {
                    "summary": "Update Task - Employee token Access Denied",
                    "value": {
                      "message": "Access denied. You do not have permission to access this resource.",
                      "status": 403,
                      "timestamp": "2026-08-07T11:57:57.93571"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Task or assigned employee not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "updateTaskUserNotFound": {
                    "summary": "Update Task - User Not Found",
                    "value": {
                      "message": "User not found",
                      "status": 404,
                      "timestamp": "2026-08-07T11:58:43.564154"
                    }
                  },
                  "updateTaskTaskNotFound": {
                    "summary": "Update Task - Task Not Found",
                    "value": {
                      "message": "Task not found",
                      "status": 404,
                      "timestamp": "2026-08-07T11:58:55.533559"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete task",
        "tags": [
          "task-controller"
        ],
        "operationId": "deleteTask",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                },
                "examples": {
                  "deleteTaskSuccessAdminToken": {
                    "summary": "Delete Task - Success Admin token",
                    "value": {
                      "message": "Task deleted successfully",
                      "timestamp": "2026-08-07T12:43:20"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "deleteTaskNoToken": {
                    "summary": "Delete Task - No token",
                    "value": {
                      "message": "Authentication failed",
                      "status": 401,
                      "timestamp": "2026-08-07T12:48:22.056422"
                    }
                  },
                  "deleteTaskInvalidToken": {
                    "summary": "Delete Task - Invalid token",
                    "value": {
                      "message": "Authentication failed",
                      "status": 401,
                      "timestamp": "2026-08-07T12:46:48.827212"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Requires ADMIN role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "deleteTaskEmployeeTokenAccessDenied": {
                    "summary": "Delete Task - Employee token Access Denied",
                    "value": {
                      "message": "Access denied. You do not have permission to access this resource.",
                      "status": 403,
                      "timestamp": "2026-08-07T12:46:11.462851"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Task not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "deleteTaskTaskNotFound": {
                    "summary": "Delete Task - Task Not Found",
                    "value": {
                      "message": "Task not found",
                      "status": 404,
                      "timestamp": "2026-08-07T12:44:43.081885"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{id}/status": {
      "put": {
        "summary": "Update task status",
        "tags": [
          "task-controller"
        ],
        "operationId": "updateTaskStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                },
                "examples": {
                  "updateTaskStatusSuccessEmployeeToken": {
                    "summary": "Update Task Status - Success Employee token",
                    "value": {
                      "id": 5,
                      "title": "Implement Frontend React Welcome Page",
                      "description": "Create Footer and Header",
                      "status": "IN_PROGRESS",
                      "priority": "MEDIUM",
                      "dueDate": "2026-08-10",
                      "assignedUserId": 19,
                      "assignedUserName": "taqu"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid status enum value",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "updateTaskStatusInvalidStatus": {
                    "summary": "Update Task Status - Invalid Status",
                    "value": {
                      "message": "Invalid status. Allowed values: PENDING, IN_PROGRESS, COMPLETED",
                      "status": 400,
                      "timestamp": "2026-08-07T17:14:35.817284"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "updateTaskStatusNoToken": {
                    "summary": "Update Task Status - No Token",
                    "value": {
                      "message": "Authentication failed",
                      "status": 401,
                      "timestamp": "2026-08-07T17:17:19.12418"
                    }
                  },
                  "updateTaskStatusSuccessEmployeeToken": {
                    "summary": "Update Task Status - Success Employee token",
                    "value": {
                      "message": "You are not authorized to update this task",
                      "status": 401,
                      "timestamp": "2026-08-07T17:10:33.24749"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Requires EMPLOYEE role or access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "updateTaskStatusAdminToken": {
                    "summary": "Update Task Status - Admin Token",
                    "value": {
                      "message": "Access denied. You do not have permission to access this resource.",
                      "status": 403,
                      "timestamp": "2026-08-07T17:16:25.75252"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Task not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "updateTaskStatusTaskNotFound": {
                    "summary": "Update Task Status - Task Not Found",
                    "value": {
                      "message": "Task not found",
                      "status": 404,
                      "timestamp": "2026-08-07T17:09:05.338205"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tasks": {
      "get": {
        "summary": "Get all tasks",
        "tags": [
          "task-controller"
        ],
        "operationId": "getAllTasks",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskResponse"
                  }
                },
                "examples": {
                  "viewAllTaskSuccessAdminToken": {
                    "summary": "View All Task - Success  Admin token",
                    "value": [
                      {
                        "id": 1,
                        "title": "Implement Login API",
                        "description": "Create login endpoint using JWT",
                        "status": "PENDING",
                        "priority": "HIGH",
                        "dueDate": "2026-08-10",
                        "assignedUserId": 17,
                        "assignedUserName": "zohaib"
                      }
                    ]
                  },
                  "viewAllTasksSuccessAdminTokenNoTasks": {
                    "summary": "View All Tasks - Success Admin token No Tasks",
                    "value": []
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "viewAllTasksNoToken": {
                    "summary": "View All Tasks - No token",
                    "value": {
                      "message": "Authentication failed",
                      "status": 401,
                      "timestamp": "2026-08-07T10:14:22.04642"
                    }
                  },
                  "viewAllTasksInvalidToken": {
                    "summary": "View All Tasks - Invalid token",
                    "value": {
                      "message": "Authentication failed",
                      "status": 401,
                      "timestamp": "2026-08-07T10:12:39.233898"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Requires ADMIN role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "viewAllTasksEmployeeTokenAccessDenied": {
                    "summary": "View All Tasks - Employee token Access Denied",
                    "value": {
                      "message": "Access denied. You do not have permission to access this resource.",
                      "status": 403,
                      "timestamp": "2026-08-07T10:10:04.673576"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new task",
        "tags": [
          "task-controller"
        ],
        "operationId": "createTask",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                },
                "examples": {
                  "createTaskByAdminTokenSuccess": {
                    "summary": "Create Task - By Admin(token) Success",
                    "value": {
                      "id": 1,
                      "title": "Implement Login API",
                      "description": "Create login endpoint using JWT",
                      "status": "PENDING",
                      "priority": "HIGH",
                      "dueDate": "2026-08-10",
                      "assignedUserId": 17,
                      "assignedUserName": "zohaib"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Validation error or invalid priority/status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "createTaskEmptyTitle": {
                    "summary": "Create Task - Empty Title",
                    "value": {
                      "message": "Title cannot be empty",
                      "status": 400,
                      "timestamp": "2026-08-06T23:23:45.960045"
                    }
                  },
                  "createTaskEmptyDescription": {
                    "summary": "Create Task - Empty Description",
                    "value": {
                      "message": "Title cannot be empty",
                      "status": 400,
                      "timestamp": "2026-08-06T23:29:00.721133"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "createTaskNoToken": {
                    "summary": "Create Task - No Token",
                    "value": {
                      "message": "Authentication failed",
                      "status": 401,
                      "timestamp": "2026-08-06T23:03:19.505755"
                    }
                  },
                  "createTaskInvalidToken": {
                    "summary": "Create Task - Invalid Token",
                    "value": {
                      "message": "Authentication failed",
                      "status": 401,
                      "timestamp": "2026-08-06T23:02:28.529716"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Requires ADMIN role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "createTaskByEmployeeTokenAccessDenied": {
                    "summary": "Create Task - By Employee(token) Access Denied",
                    "value": {
                      "message": "Access denied. You do not have permission to access this resource.",
                      "status": 403,
                      "timestamp": "2026-08-06T23:01:13.413435"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Assigned employee not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "createTaskUserNotFound": {
                    "summary": "Create Task - Assigned User Not Found",
                    "value": {
                      "message": "User not found",
                      "status": 404,
                      "timestamp": "2026-08-06T23:21:12.182218"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict - Resource conflict or user not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "createTaskInvalidUserId": {
                    "summary": "Create Task - Invalid User ID",
                    "value": {
                      "message": "User not found",
                      "status": 409,
                      "timestamp": "2026-08-06T23:21:12.182218"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tasks/my-tasks": {
      "get": {
        "summary": "Get tasks assigned to current user",
        "tags": [
          "task-controller"
        ],
        "operationId": "getMyTasks",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskResponse"
                  }
                },
                "examples": {
                  "viewMyTasksSuccessEmployeeToken": {
                    "summary": "View My Tasks - Success Employee token",
                    "value": [
                      {
                        "id": 4,
                        "title": "Implement Login API",
                        "description": "Create login endpoint using JWT",
                        "status": "PENDING",
                        "priority": "HIGH",
                        "dueDate": "2026-08-10",
                        "assignedUserId": 17,
                        "assignedUserName": "zohaib"
                      }
                    ]
                  },
                  "viewMyTasksSuccessEmployeeTokenNoTasks": {
                    "summary": "View My Tasks - Success Employee token No Tasks",
                    "value": []
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Missing or invalid JWT token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "viewMyTasksNoToken": {
                    "summary": "View My Tasks - No token",
                    "value": {
                      "message": "Authentication failed",
                      "status": 401,
                      "timestamp": "2026-08-07T16:17:27.466374"
                    }
                  },
                  "viewMyTasksInvalidToken": {
                    "summary": "View My Tasks - Invalid token",
                    "value": {
                      "message": "Authentication failed",
                      "status": 401,
                      "timestamp": "2026-08-07T16:16:12.874565"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Requires EMPLOYEE role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "viewMyTasksAdminTokenAccessDenied": {
                    "summary": "View My Tasks - Admin token Access Denied",
                    "value": {
                      "message": "Access denied. You do not have permission to access this resource.",
                      "status": 403,
                      "timestamp": "2026-08-07T16:15:24.423722"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "TaskRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "minLength": 1
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ]
          },
          "dueDate": {
            "type": "string",
            "format": "date"
          },
          "assignedUserId": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "assignedUserId",
          "description",
          "dueDate",
          "priority",
          "title"
        ]
      },
      "TaskResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "priority": {
            "type": "string"
          },
          "dueDate": {
            "type": "string",
            "format": "date"
          },
          "assignedUserId": {
            "type": "integer",
            "format": "int64"
          },
          "assignedUserName": {
            "type": "string"
          },
          "assignedTo": {
            "type": "string"
          },
          "createdAT": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UpdateTaskStatusRequest": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "TODO",
              "IN_PROGRESS",
              "DONE"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
      "DeleteResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  }
}